# openapi.yaml
openapi: 3.0.3
info:
  title: 邻里知光 API
  version: 1.0.0
  description: |
    基于《邻里知光-技术分档-详细版》与补充 docx 整理的目标态接口文档。
    当前工作区无代码快照，本规范用于联调、评审与后续 Swagger/Knife4j/Postman 导入。
servers:
  - url: /api/v1
    description: 默认 API 前缀
security:
  - BearerAuth: []
...
paths:
...
  /profile/me:
    get:
      tags: [Profile]
      summary: 查询当前登录用户资料
      operationId: getMyProfile
      x-maturity: 已实现
      x-source: 当前代码
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileData'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags: [Profile]
      summary: 局部更新当前登录用户资料
      operationId: patchMyProfile
      x-maturity: 已实现
      x-source: 当前代码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfilePatchRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileData'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /profile/avatar:
    post:
      tags: [Profile]
      summary: 单独更新当前登录用户头像
      operationId: updateProfileAvatar
      x-maturity: 已实现
      x-source: 当前代码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileAvatarRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileData'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /profile/users/{userId}:
    get:
      tags: [Profile]
      summary: 查询指定用户主页资料
      operationId: getUserProfile
      security: []
      x-maturity: 已实现
      x-source: 当前代码
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileData'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /profile/users/{userId}/posts:
    get:
      tags: [Profile]
      summary: 查询指定用户主页可见的发布内容
      operationId: getUserProfilePosts
      security: []
      x-maturity: 已实现
      x-source: 当前代码
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PostPageData'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /profile/users/{userId}/following:
    get:
      tags: [Profile]
      summary: 查询指定用户的关注列表资料视图
      operationId: getProfileFollowing
      security: []
      x-maturity: 已实现
      x-source: 当前代码
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileListData'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /profile/users/{userId}/followers:
    get:
      tags: [Profile]
      summary: 查询指定用户的粉丝列表资料视图
      operationId: getProfileFollowers
      security: []
      x-maturity: 已实现
      x-source: 当前代码
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileListData'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /storage/presign:
    post:
      tags: [Storage]
      summary: 申请 OSS 预签名上传地址
      operationId: createPresign
      x-maturity: 已实现
      x-source: 基于需求推导
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignRequest'
      responses:
        '200':
          description: 生成成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PresignData'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
...